% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pdsi.R
\name{read_pdsi_input}
\alias{read_pdsi_input}
\title{Read the input data of v2003 scPDSI to R}
\usage{
read_pdsi_input(file)
}
\arguments{
\item{file}{Character, input file path.}
}
\description{
Read the input data of v2003 scPDSI to R
}
\keyword{internal}
