% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pdsi.R
\name{read_pdsi_ouput}
\alias{read_pdsi_ouput}
\title{Read the output data of v2003 scPDSI to R}
\usage{
read_pdsi_ouput(outdir, scale = "monthly")
}
\arguments{
\item{outdir}{Directory of output data.}

\item{scale}{Character, \code{monthly} or \code{weekly}.}
}
\description{
Read the output data of v2003 scPDSI to R
}
\keyword{internal}
